refactor: reorganize package structure and update imports
- Consolidated package naming for consistency:
  - Moved `api.template` to `common` in shared classes.
  - Updated `api.template` to `tests.petstore.api` in Petstore-specific files.
- Adjusted imports across all affected files to reflect new package structure.
- Minor update to `README.md` to improve documentation organization.
- Downgraded Allure version in `build.gradle` for better compatibility.
