--- .github/workflows/api-tests.yml
@@ -1,57 +1,0 @@
-name: API Tests
-
-on:
-  push:
-    branches: [ main, master ]
-  pull_request:
-    branches: [ main, master ]
-  workflow_dispatch:
-
-permissions:
-  contents: read
-  issues: read
-  checks: write
-  pull-requests: write
-
-jobs:
-  api-tests:
-    runs-on: ubuntu-latest
-    
-    steps:
-      - uses: actions/checkout@v3
-
-      - name: Set up JDK 17
-        uses: actions/setup-java@v3
-        with:
-          java-version: '17'
-          distribution: 'temurin'
-          cache: gradle
-
-      - name: ⚙ Grant execute permission for gradlew
-        run: chmod +x gradlew
-
-      - name: Build with Gradle
-        run: ./gradlew build -x spotlessCheck -x test
-
-      - name: Check Code Format
-        run: ./gradlew spotlessCheck
-
-      - name: Run API Tests
-        run: ./gradlew apiTests
-
-      #      - name: Generate Allure Report
-      #        if: always()
-      #        run: ./gradlew allureReport
-      #
-      #      - name: Upload Allure Report
-      #        if: always()
-      #        uses: actions/upload-artifact@v4
-      #        with:
-      #          name: allure-report
-      #          path: build/reports/allure-report
-
-      - name: Publish API Test Results
-        if: (!cancelled())
-        uses: EnricoMi/publish-unit-test-result-action@v2
-        with:
-          files: "*/build/test-results/test/TEST-*.xml"
