--- common/src/main/java/com/github/artbi/api/template/conditions/Condition.java
@@ -1,13 +1,0 @@
-/*
- * Copyright (c) 2025
- */
-package com.github.artbi.api.template.conditions;
-
-import io.restassured.response.Response;
-
-public interface Condition {
-    void check(Response response);
-
-    @Override
-    String toString();
-}
