--- common/src/main/java/com/github/artbi/api/template/conditions/ContentTypeCondition.java
@@ -1,23 +1,0 @@
-/*
- * Copyright (c) 2025
- */
-package com.github.artbi.api.template.conditions;
-
-import io.restassured.response.Response;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-public class ContentTypeCondition implements Condition {
-
-    private final String contentType;
-
-    @Override
-    public void check(Response response) {
-        response.then().assertThat().contentType(contentType);
-    }
-
-    @Override
-    public String toString() {
-        return "content type is [%s]".formatted(contentType);
-    }
-}
