--- common/src/main/java/com/github/artbi/api/template/config/BaseConfig.java
@@ -1,22 +1,0 @@
-/*
- * Copyright (c) 2025
- */
-package com.github.artbi.api.template.config;
-
-import org.aeonbits.owner.Config;
-import org.aeonbits.owner.ConfigFactory;
-
-@Config.Sources({"classpath:config.properties"})
-public interface BaseConfig extends Config {
-
-  static BaseConfig get() {
-    return ConfigFactory.create(BaseConfig.class, System.getProperties());
-  }
-
-  @Key("api.base.url")
-  String apiBaseUrl();
-
-  @Key("logging.enabled")
-  @DefaultValue("false")
-  boolean loggingEnabled();
-}
