+++ common/src/main/java/com/github/artbi/common/conditions/Condition.java
@@ -1,0 +1,13 @@
+/*
+ * Copyright (c) 2025
+ */
+package com.github.artbi.common.conditions;
+
+import io.restassured.response.Response;
+
+public interface Condition {
+    void check(Response response);
+
+    @Override
+    String toString();
+}
