+++ common/src/main/java/com/github/artbi/common/conditions/Conditions.java
@@ -1,0 +1,23 @@
+/*
+ * Copyright (c) 2025
+ */
+package com.github.artbi.common.conditions;
+
+import lombok.experimental.UtilityClass;
+import org.hamcrest.Matcher;
+
+@UtilityClass
+public class Conditions {
+
+    public static StatusCodeCondition statusCode(int code) {
+        return new StatusCodeCondition(code);
+    }
+
+    public static ContentTypeCondition contentType(String contentType) {
+        return new ContentTypeCondition(contentType);
+    }
+
+    public static BodyFieldCondition bodyField(String jsonPath, Matcher matcher) {
+        return new BodyFieldCondition(jsonPath, matcher);
+    }
+}
