+++ common/src/main/java/com/github/artbi/common/conditions/StatusCodeCondition.java
@@ -1,0 +1,23 @@
+/*
+ * Copyright (c) 2025
+ */
+package com.github.artbi.common.conditions;
+
+import io.restassured.response.Response;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public class StatusCodeCondition implements Condition {
+
+    private final int statusCode;
+
+    @Override
+    public void check(Response response) {
+        response.then().assertThat().statusCode(statusCode);
+    }
+
+    @Override
+    public String toString() {
+        return "status code is %d".formatted(statusCode);
+    }
+}
