+++ petstore-api-tests/src/test/java/com/github/artbi/api/petstore/data/generators/PetDataGenerator.java
@@ -1,0 +1,7 @@
+/*
+ * Copyright (c) 2025
+ */
+package com.github.artbi.api.petstore.data.generators;
+
+public class PetDataGenerator {
+}
