--- petstore-api-tests/src/test/java/com/github/artbi/api/template/data/generators/PetDataGenerator.java
@@ -1,7 +1,0 @@
-/*
- * Copyright (c) 2025
- */
-package com.github.artbi.api.template.data.generators;
-
-public class PetDataGenerator {
-}
